refactor: 모듈별 폴더 분리
