+++ ai-server/constants.py
@@ -4,5 +4,9 @@ CONVERSATION_ID = {
     "EXAMPLE_4" : 172,
     "EXAMPLE_5" : 176,
     "EXAMPLE_10" : 178,
-    "EXAMPLE_11" : 179
+    "EXAMPLE_11" : 179,
+    "EXAMPLE_1_KOR" : 146,
+    "EXAMPLE_2_KOR" : 152,
+    "EXAMPLE_5_KOR" : 162,
+    "EXAMPLE_6_KOR" : 153
 }
